update rental and student
