+++ student.rb
@@ -1,8 +1,6 @@
-require './person'
+require_relative 'person'
 
 class Student < Person
-  attr_accessor :classroom
-
   def initialize(id, age, name: 'Unknown', parent_permission: true, classroom: 'Unknown')
     super(id, age, name: name, parent_permission: parent_permission)
     @classroom = classroom
@@ -11,10 +9,9 @@ class Student < Person
   def play_hooky
     '¯\(ツ)/¯'
   end
-end
 
-student = Student.new(1, 16, name: 'Alice', parent_permission: false, classroom: 'High School')
-puts student.name # Output: Alice
-puts student.age # Output: 16
-puts student.classroom # Output: High School
-puts student.can_use_services? # Output: true (inherits from Person class)
+  def classroom=(classroom)
+    @classroom = classroom
+    classroom.classroom_students.push(self) unless classroom.students.include?(self)
+  end
+end
